adding score data to match table db
