+++ server/controllers/utils/matchUtils.controller.ts
@@ -1,0 +1,52 @@
+import { FastifyRequest, FastifyReply } from 'fastify';
+
+export async function getStats(request: FastifyRequest, reply: FastifyReply) {
+	try {
+		const { username } = request.params as { username: string };
+		const user = await request.server.prisma.user.findUnique({
+			where: { username },
+			select: {
+				wins: true, 
+				losses: true,
+			}
+		});
+
+	return reply.code(200).send(user); // check when/how this is being used and therefore what to return
+	} catch (error) {
+		return reply.status(500).send({ error: 'Failed to fetch user statistics' });
+	}
+};
+
+export async function updateWins(userId: number, request: FastifyRequest, reply: FastifyReply) {
+	try {
+		const userID = Number(userId);
+		await request.server.prisma.user.update({
+			where: { id: userID },
+			data: { 
+				wins: {
+					increment: 1,
+				}
+			}
+		});
+		reply.code(200);
+	} catch (error) {
+		return reply.status(500).send({ error: 'Failed to update user win count' });
+	}
+};
+
+export async function updateLosses(userId: number, request: FastifyRequest, reply: FastifyReply) {
+	try {
+		const userID = Number(userId);
+		await request.server.prisma.user.update({
+			where: { id: userID },
+			data: { 
+				losses: {
+					increment: 1,
+				}
+			}
+		});
+		reply.code(200);
+	} catch (error) {
+		return reply.status(500).send({ error: 'Failed to update user loss count' });
+	}
+};
